moved query_interface to protocols directory
